Moved non-reusable components to modules
