+++ src/modules/notes/Note.tsx
