+++ src/modules/notes/NotesProgressBar.tsx
