+++ src/modules/notes/SearchBar.tsx
